requirements.txt
dj-database-url==0.4.0
Django==1.8.8
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-bootstrap3
django-widget-tweaks
pycrypto